Discord/AaronsDiscordProject/requirements.txt
intents~=0.3.0
python-dotenv~=1.0.0